Player works + Display works!
The player is in the perfect spot: at the bottom of the page, and does not load in with the other results. Consequently, each songcard will no longer be rendered as its own component. This made it complicated to grab the uri from each song to play. It was a trade between functional player but unappealing display, or vice versa.